DDP-5802 - fix unit tests
