+++ pepper-apis/src/main/java/org/broadinstitute/ddp/service/actvityinstancebuilder/context/AIBuilderContext.java
@@ -50,12 +50,8 @@ public class AIBuilderContext {
     public AIBuilderContext(Handle handle, AIBuilderParams params) {
         this.handle = handle;
         this.params = params;
-
         LanguageDto languageDto = LanguageStore.get(params.getIsoLangCode());
-        if (languageDto == null) {
-            throw new IllegalStateException("Invalid language specified: " + params.getIsoLangCode());
-        }
-        this.langCodeId = languageDto.getId();
+        this.langCodeId = languageDto != null ? languageDto.getId() : 0;
     }
 
     public AIBuildStep getBuildStep() {
